+++ SensorVoltage.h
@@ -29,13 +29,13 @@ You should have received a copy of the GNU General Public License along with thi
 class SensorVoltage : public ISensor
 {
 public:
-	 SensorVoltage(const char * name, int analogPin, long resistorToGround, long resistorToVoltage);
+	 SensorVoltage(const char * name, uint8_t analogPin, long resistorToGround, long resistorToVoltage);
 	 sensorVal_t getValueInt(); // return in mV
 	 float getValueFloat(); // return in V
 	 int getSamplingTime();
 
 private:
-	int mPin;
+	uint8_t mPin;
 	float mDivider;
 };
